Reintroduced api types as being part of api.svelte.ts
